refactor(develop): correct classes name and pawn representation
