+++ chess-pieces/types/index.ts
@@ -1,5 +1,6 @@
 export * from "./color"
 export * from "./chess-board"
 export * from "./piece"
-export * from "./pieces-factory"
+export * from "./piecesFactory"
 export * from "./pieceType"
+export * from "./playerMove"
