+++ chess-pieces/types/piecesFactory.ts
